pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "codetracer"
version = "0.2.0"
description = "CodeTracer: self-evolving agent trajectory diagnosis & replay system"
requires-python = ">=3.10"

dependencies = [
    "pyyaml",
    "jinja2",
    "pydantic>=2.0",
    "tenacity",
    "rich",
    "typer",
    "python-dotenv",
    "platformdirs",
    "openai",
    "requests",
    "datasets",
    "huggingface_hub",
    "inquirer",
    "readchar",
]

[project.optional-dependencies]
docker = [
    "docker",
]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
codetracer = "codetracer.cli:app"
codetracer-batch = "codetracer.scripts.batch_run:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["codetracer*"]

[tool.setuptools.package-data]
codetracer = [
    "config/**/*",
    "skills/seed/**/*",
    "prompts/presets/**/*",
    "benches/seed/**/*",
    "templates/presets/**/*",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501", "E402", "E722"]
